refactor : 출입국 폼 fxml로 적용
