+++ src/main/resources/org/example/onesteponestamp/fxml/ImmigrationFormView.fxml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.GridPane?>
+<GridPane fx:id="form" alignment="CENTER" hgap="10" vgap="10" xmlns="http://javafx.com/javafx/17"
+  xmlns:fx="http://javafx.com/fxml/1"
+  fx:controller="org.example.onesteponestamp.javafx.ImmigrationFormController">
+  <padding>
+    <Insets top="10" right="10" bottom="10" left="10"/>
+  </padding>
+  <Label text="자동출입국신청번호" GridPane.columnIndex="0" GridPane.rowIndex="0"
+    style="-fx-font-weight: bold; -fx-font-size: 16px;"/>
+  <TextField fx:id="applyNoField" GridPane.columnIndex="1" GridPane.rowIndex="0"/>
+  <Button text="제출" onAction="#verificationForm" GridPane.columnIndex="2" GridPane.rowIndex="0"/>
+</GridPane>
